refactor(comp: form): add useFormSize
