+++ packages/components/form/index.ts
@@ -18,7 +18,6 @@ const IxFormWrapper = FormWrapper as unknown as FormWrapperComponent
 export { IxForm, IxFormItem, IxFormWrapper }
 
 export { FORM_TOKEN, FORM_ITEM_TOKEN, useFormItemRegister } from './src/token'
-export type { FormContext } from './src/token'
 
 export type {
   FormInstance,
@@ -36,3 +35,5 @@ export type {
   FormValidateMessage,
   FormSize,
 } from './src/types'
+
+export type { FormContext } from './src/token'
